expr: make `PropertizedDatum` available to other crates
